+++ 06_generate_ubuntu_console_root.sh
@@ -32,6 +32,7 @@ if [ -d ./ignore/.root ] ; then
 fi
 mkdir -p ./ignore/.root
 
+echo "Extracting: ubuntu-23.04-console-riscv64-${datestamp}/riscv64-rootfs-ubuntu-lunar.tar"
 tar xfp ./deploy/ubuntu-23.04-console-riscv64-${datestamp}/riscv64-rootfs-ubuntu-lunar.tar -C ./ignore/.root
 sync
 
@@ -46,7 +47,19 @@ cd ./ignore/.root/
 ln -L -f -s -v /lib/systemd/system/resize_filesystem.service --target-directory=./etc/systemd/system/multi-user.target.wants/
 cd ../../
 
-cp -v ./ignore/.root/etc/bbb.io/templates/eth0-DHCP.network ./ignore/.root/etc/systemd/network/eth0.network || true
+if [ -f ./ignore/.root/etc/bbb.io/templates/eth0-DHCP.network ] ; then
+	cp -v ./ignore/.root/etc/bbb.io/templates/eth0-DHCP.network ./ignore/.root/etc/systemd/network/eth0.network || true
+else
+	echo '[Match]' > ./ignore/.root/etc/systemd/network/eth0.network
+	echo 'Name=eth0' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo 'Type=ether' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo '' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo '[Link]' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo 'RequiredForOnline=yes' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo '' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo '[Network]' >> ./ignore/.root/etc/systemd/network/eth0.network
+	echo 'DHCP=ipv4' >> ./ignore/.root/etc/systemd/network/eth0.network
+fi
 
 #Cleanup large firmware's..
 rm -rf ./ignore/.root/usr/lib/firmware/amdgpu/ || true
